requirements.txt
mediapipe==0.10.33
moviepy==1.0.3
numpy==2.4.4
protobuf==7.34.1
Requests==2.33.1
openai-whisper
yt_dlp==2026.3.17
opencv-python
google-genai
google-api-python-client
google-auth-oauthlib
google-auth-httplib2